feat: move arrays to data folder, update imports
