+++ lab_3/react-ts-template/src/data/CharactersArray.ts
@@ -1,0 +1,44 @@
+import { DetailPropType } from 'types/Detail';
+
+export const CharactersArr: DetailPropType[] = [
+  {
+    name: 'DOCTOR STRANGE',
+    desc: 'Possessing vast magical knowledge and skills to call upon, he now serves as Earth’s newest Master of the Mystic Arts.',
+    img: 'https://i.annihil.us/u/prod/marvel/i/mg/9/f0/57fe4ef408a29/standard_incredible.jpg',
+    id: '1',
+    comics: ['1', '2'],
+    series: ['1', '2']
+  },
+  {
+    name: 'Thor',
+    desc: 'Possessing vast magical knowledge and skills to call upon, he now serves as Earth’s newest Master of the Mystic Arts.',
+    img: 'https://i.annihil.us/u/prod/marvel/i/mg/9/f0/57fe4ef408a29/standard_incredible.jpg',
+    id: '2',
+    comics: ['2', '5'],
+    series: ['1', '3']
+  },
+  {
+    name: 'Iron Man',
+    desc: 'Possessing vast magical knowledge and skills to call upon, he now serves as Earth’s newest Master of the Mystic Arts.',
+    img: 'https://i.annihil.us/u/prod/marvel/i/mg/9/f0/57fe4ef408a29/standard_incredible.jpg',
+    id: '3',
+    comics: ['3', '4'],
+    series: ['3', '5']
+  },
+  {
+    name: 'Spider man',
+    desc: 'Possessing vast magical knowledge and skills to call upon, he now serves as Earth’s newest Master of the Mystic Arts.',
+    img: 'https://i.annihil.us/u/prod/marvel/i/mg/9/f0/57fe4ef408a29/standard_incredible.jpg',
+    id: '4',
+    comics: ['5', '6'],
+    series: ['5', '6']
+  },
+  {
+    name: 'DOCTOR',
+    desc: 'Possessing vast magical knowledge and skills to call upon, he now serves as Earth’s newest Master of the Mystic Arts.',
+    img: 'https://i.annihil.us/u/prod/marvel/i/mg/9/f0/57fe4ef408a29/standard_incredible.jpg',
+    id: '5',
+    comics: ['3', '4'],
+    series: ['3', '4']
+  }
+];
