+++ lab_3/react-ts-template/src/data/ComicsArray.ts
@@ -1,0 +1,46 @@
+import { DetailPropType } from 'types/Detail';
+
+export const ComicsArr: DetailPropType[] = [
+  {
+    name: 'New Mutants: Facsimile Edition (2022) #98',
+    desc: "The dynamic debut of Deadpool! It's the beginning of the end for the New Mutants as Cable continues his reinvention of the squad into his own paramilitary force!",
+    img: 'https://i.annihil.us/u/prod/marvel/i/mg/9/50/632cd36b4da77/portrait_uncanny.jpg',
+    id: '1',
+    characters: ['1', '2', '3']
+  },
+  {
+    name: 'Spider Man',
+    desc: "The dynamic debut of Deadpool! It's the beginning of the end for the New Mutants as Cable continues his reinvention of the squad into his own paramilitary force!",
+    img: 'https://i.annihil.us/u/prod/marvel/i/mg/d/60/63658cf8efef2/portrait_uncanny.jpg',
+    id: '2',
+    characters: ['3', '1', '5']
+  },
+  {
+    name: 'Thor',
+    desc: "The dynamic debut of Deadpool! It's the beginning of the end for the New Mutants as Cable continues his reinvention of the squad into his own paramilitary force!",
+    img: 'https://i.annihil.us/u/prod/marvel/i/mg/9/50/632cd36b4da77/portrait_uncanny.jpg',
+    id: '3',
+    characters: ['2', '3', '4']
+  },
+  {
+    name: 'Iron Man #1',
+    desc: "The dynamic debut of Deadpool! It's the beginning of the end for the New Mutants as Cable continues his reinvention of the squad into his own paramilitary force!",
+    img: 'https://i.annihil.us/u/prod/marvel/i/mg/d/60/63658cf8efef2/portrait_uncanny.jpg',
+    id: '4',
+    characters: ['1', '2', '3']
+  },
+  {
+    name: 'Iron Man #15',
+    desc: "The dynamic debut of Deadpool! It's the beginning of the end for the New Mutants as Cable continues his reinvention of the squad into his own paramilitary force!",
+    img: 'https://i.annihil.us/u/prod/marvel/i/mg/9/50/632cd36b4da77/portrait_uncanny.jpg',
+    id: '5',
+    characters: ['2', '5', '2']
+  },
+  {
+    name: 'Clay Man #994',
+    desc: "The dynamic debut of Deadpool! It's the beginning of the end for the New Mutants as Cable continues his reinvention of the squad into his own paramilitary force!",
+    img: 'https://i.annihil.us/u/prod/marvel/i/mg/d/60/63658cf8efef2/portrait_uncanny.jpg',
+    id: '6',
+    characters: ['1', '3', '5']
+  }
+];
