+++ lab_3/react-ts-template/src/pages/Comics/ComicsDetails.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import Detail from 'components/Detail';
 import { useParams } from 'react-router-dom';
-import { comicsArr } from './Comics';
+import { ComicsArr } from 'data/ComicsArray';
 
 const ComicsDetails: React.FC = () => {
   const { id } = useParams();
 
-  return <Detail {...comicsArr.filter((e) => e.id === id)[0]} />;
+  return <Detail {...ComicsArr.filter((e) => e.id === id)[0]} />;
 };
 
 export default ComicsDetails;
