+++ lab_3/react-ts-template/src/pages/Series/SeriesDetails.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import Detail from 'components/Detail';
 import { useParams } from 'react-router-dom';
-import { seriesArr } from './Series';
+import { SeriesArr } from 'data/SeriesArray';
 
 const SeriesDetails: React.FC = () => {
   const { id } = useParams();
 
-  return <Detail {...seriesArr.filter((e) => e.id === id)[0]} />;
+  return <Detail {...SeriesArr.filter((e) => e.id === id)[0]} />;
 };
 
 export default SeriesDetails;
